Extracted out targetdb converted suites in a separate package, Using the schemaRegistry's Schema details for passing to control file of sqlldr for datatypes
