+++ yb-voyager/src/tgtdb/oracle.go
@@ -19,10 +19,7 @@ import (
 	"bufio"
 	"context"
 	"database/sql"
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
-	"math"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -35,6 +32,7 @@ import (
 	log "github.com/sirupsen/logrus"
 	"github.com/yugabyte/yb-voyager/yb-voyager/src/sqlldr"
 	"github.com/yugabyte/yb-voyager/yb-voyager/src/utils"
+	"github.com/yugabyte/yb-voyager/yb-voyager/src/tgtdb/suites"
 )
 
 
@@ -43,186 +41,6 @@ type TargetOracleDB struct {
 	tconf *TargetConf
 	oraDB *sql.DB
 	conn  *sql.Conn
-}
-
-var oraValueConverterSuite = map[string]ConverterFn{
-	"DATE": func(columnValue string, formatIfRequired bool) (string, error) {
-		epochMilliSecs, err := strconv.ParseInt(columnValue, 10, 64) // from oracle for DATE type debezium gives epoch milliseconds or type is io.debezium.time.Timestamp
-		if err != nil {
-			return columnValue, fmt.Errorf("parsing epoch milliseconds: %v", err)
-		}
-		epochSecs := epochMilliSecs / 1000
-		parsedTime := time.Unix(epochSecs, 0).Local()
-		oracleDateFormat := "02-01-2006" //format: DD-MON-YYYY
-		formattedDate := parsedTime.Format(oracleDateFormat)
-		if err != nil {
-			return "", fmt.Errorf("parsing date: %v", err)
-		}
-		if formatIfRequired {
-			formattedDate = fmt.Sprintf("'%s'", formattedDate)
-		}
-		return formattedDate, nil
-	}, // TODO: change type to DATE as per typename in db
-	"io.debezium.time.Timestamp": func(columnValue string, formatIfRequired bool) (string, error) {
-		epochMilliSecs, err := strconv.ParseInt(columnValue, 10, 64)
-		if err != nil {
-			return columnValue, fmt.Errorf("parsing epoch milliseconds: %v", err)
-		}
-		epochSecs := epochMilliSecs / 1000
-		timestamp := time.Unix(epochSecs, 0).Local()
-		oracleTimestampFormat := "02-01-2006 03.04.05.000 PM" //format: DD-MM-YY HH.MI.SS.FFF PM
-		formattedTimestamp := timestamp.Format(oracleTimestampFormat)
-		if formatIfRequired {
-			formattedTimestamp = fmt.Sprintf("'%s'", formattedTimestamp)
-		}
-		return formattedTimestamp, nil
-	},
-	"io.debezium.time.MicroTimestamp": func(columnValue string, formatIfRequired bool) (string, error) {
-		epochMicroSecs, err := strconv.ParseInt(columnValue, 10, 64)
-		if err != nil {
-			return columnValue, fmt.Errorf("parsing epoch microseconds: %v", err)
-		}
-		epochSeconds := epochMicroSecs / 1000000
-		epochNanos := (epochMicroSecs % 1000000) * 1000
-		microTimeStamp := time.Unix(epochSeconds, epochNanos).Local()
-		oracleTimestampFormat := "02-01-2006 03.04.05.000000 PM" //format: DD-MON-YYYY HH.MI.SS.FFFFFF PM
-		formattedTimestamp := microTimeStamp.Format(oracleTimestampFormat)
-		if formatIfRequired {
-			formattedTimestamp = fmt.Sprintf("'%s'", formattedTimestamp)
-		}
-		return formattedTimestamp, nil
-	},
-	"io.debezium.time.NanoTimestamp": func(columnValue string, formatIfRequired bool) (string, error) {
-		epochNanoSecs, err := strconv.ParseInt(columnValue, 10, 64)
-		if err != nil {
-			return "", fmt.Errorf("parsing epoch nanoseconds: %v", err)
-		}
-		epochSeconds := epochNanoSecs / 1000000000
-		epochNanos := epochNanoSecs % 1000000000
-		nanoTimestamp := time.Unix(epochSeconds, epochNanos).Local()
-		oracleTimestampFormat := "02-01-2006 03.04.05.000000000 PM" //format: DD-MON-YYYY HH.MI.SS.FFFFFFFFF PM
-		formattedTimestamp := nanoTimestamp.Format(oracleTimestampFormat)
-		if formatIfRequired {
-			formattedTimestamp = fmt.Sprintf("'%s'", formattedTimestamp)
-		}
-		return formattedTimestamp, nil
-	},
-	"io.debezium.time.ZonedTimestamp": func(columnValue string, formatIfRequired bool) (string, error) {
-		debeziumFormat := "2006-01-02T15:04:05Z07:00"
-
-		parsedTime, err := time.Parse(debeziumFormat, columnValue)
-		if err != nil {
-			return "", fmt.Errorf("parsing timestamp: %v", err)
-		}
-
-		oracleFormat := "06-01-02 3:04:05.000000000 PM -07:00"
-		if parsedTime.Location().String() == "UTC" { //LOCAL TIMEZONE Case
-			oracleFormat = "06-01-02 3:04:05.000000000 PM" // TODO: for timezone ones sqlldr is inserting as GMT though GMT/UTC is similar
-		}
-		formattedTimestamp := parsedTime.Format(oracleFormat)
-		if formatIfRequired {
-			formattedTimestamp = fmt.Sprintf("'%s'", formattedTimestamp)
-		}
-		return formattedTimestamp, nil
-	},
-	"BYTES": func(columnValue string, formatIfRequired bool) (string, error) {
-		//decode base64 string to bytes
-		decodedBytes, err := base64.StdEncoding.DecodeString(columnValue) //e.g.`////wv==` -> `[]byte{0x00, 0x00, 0x00, 0x00}`
-		if err != nil {
-			return columnValue, fmt.Errorf("decoding base64 string: %v", err)
-		}
-		//convert bytes to hex string e.g. `[]byte{0x00, 0x00, 0x00, 0x00}` -> `\\x00000000`
-		hexString := ""
-		for _, b := range decodedBytes {
-			hexString += fmt.Sprintf("%02x", b)
-		}
-		hexValue := ""
-		if formatIfRequired {
-			hexValue = fmt.Sprintf("'%s'", hexString) // in insert statement no need of escaping the backslash and add quotes
-		} else {
-			hexValue = fmt.Sprintf(`%s`, hexString) // in data file need to escape the backslash
-		}
-		return string(hexValue), nil
-	},
-	"MAP": func(columnValue string, _ bool) (string, error) {
-		mapValue := make(map[string]interface{})
-		err := json.Unmarshal([]byte(columnValue), &mapValue)
-		if err != nil {
-			return columnValue, fmt.Errorf("parsing map: %v", err)
-		}
-		var transformedMapValue string
-		for key, value := range mapValue {
-			transformedMapValue = transformedMapValue + fmt.Sprintf("\"%s\"=>\"%s\",", key, value)
-		}
-		return fmt.Sprintf("'%s'", transformedMapValue[:len(transformedMapValue)-1]), nil //remove last comma and add quotes
-	},
-	"STRING": func(columnValue string, formatIfRequired bool) (string, error) {
-		if formatIfRequired {
-			formattedColumnValue := strings.Replace(columnValue, "'", "''", -1)
-			return fmt.Sprintf("'%s'", formattedColumnValue), nil
-		} else {
-			return columnValue, nil
-		}
-	},
-	"INTERVAL YEAR TO MONTH": func(columnValue string, formatIfRequired bool) (string, error) {
-		//columnValue format: P-1Y-5M0DT0H0M0S
-		splits := strings.Split(strings.TrimPrefix(columnValue, "P"), "M") // ["-1Y-5", "0DT0H0M0S"]
-		yearsMonths := strings.Split(splits[0], "Y")                       // ["-1", "-5"]
-		years, err := strconv.ParseInt(yearsMonths[0], 10, 64)             // -1
-		if err != nil {
-			return "", fmt.Errorf("parsing years: %v", err)
-		}
-		months, err := strconv.ParseInt(yearsMonths[1], 10, 64) // -5
-		if err != nil {
-			return "", fmt.Errorf("parsing months: %v", err)
-		}
-		if years < 0 || months < 0 {
-			years = int64(math.Abs(float64(years)))
-			months = int64(math.Abs(float64(months)))
-			columnValue = fmt.Sprintf("-%d-%d", years, months) // -1-5
-		} else {
-			columnValue = fmt.Sprintf("%d-%d", years, months) // 1-5
-		}
-		if formatIfRequired {
-			columnValue = fmt.Sprintf("'%s'", columnValue)
-		}
-		return columnValue, nil
-	},
-	"INTERVAL DAY TO SECOND": func(columnValue string, formatIfRequired bool) (string, error) {
-		//columnValue format: P0Y0M24DT23H34M5.878667S //TODO check regex will be better or not
-		splits := strings.Split(strings.TrimPrefix(columnValue, "P"), "M") // ["0Y0M", "24DT23H34, 5.878667S"]
-		daysTime := strings.Split(splits[1], "DT")                         // ["24", "23H34M5.878667S"]
-		days, err := strconv.ParseInt(daysTime[0], 10, 64)                 // 24
-		if err != nil {
-			return "", fmt.Errorf("parsing days: %v", err)
-		}
-		time := strings.Split(daysTime[1], "H")         // ["23", "34"]
-		hours, err := strconv.ParseInt(time[0], 10, 64) // 23
-		if err != nil {
-			return "", fmt.Errorf("parsing hours: %v", err)
-		}
-		mins, err := strconv.ParseInt(time[1], 10, 64) // 34
-		if err != nil {
-			return "", fmt.Errorf("parsing minutes: %v", err)
-		}
-		seconds, err := strconv.ParseFloat(strings.TrimSuffix(splits[2], "S"), 64) // 5.878667
-		if err != nil {
-			return "", fmt.Errorf("parsing seconds: %v", err)
-		}
-		if days < 0 || hours < 0 || mins < 0 || seconds < 0 {
-			days = int64(math.Abs(float64(days)))
-			hours = int64(math.Abs(float64(hours)))
-			mins = int64(math.Abs(float64(mins)))
-			seconds = math.Abs(seconds)
-			columnValue = fmt.Sprintf("-%d %d:%d:%.9f", days, hours, mins, seconds) // -24 23:34:5.878667
-		} else {
-			columnValue = fmt.Sprintf("%d %d:%d:%.9f", days, hours, mins, seconds) // 24 23:34:5.878667
-		}
-		if formatIfRequired {
-			columnValue = fmt.Sprintf("'%s'", columnValue)
-		}
-		return columnValue, nil
-	},
 }
 
 var TableToColumnDataTypesCache = make(map[string]map[string]string)
@@ -498,14 +316,14 @@ func (tdb *TargetOracleDB) IsNonRetryableCopyError(err error) bool {
 	return false
 }
 
-func (tdb *TargetOracleDB) ImportBatch(batch Batch, args *ImportBatchArgs, exportDir string) (int64, error) {
+func (tdb *TargetOracleDB) ImportBatch(batch Batch, args *ImportBatchArgs, exportDir string, tableSchema map[string]map[string]string) (int64, error) {
 	tdb.Lock()
 	defer tdb.Unlock()
 
 	var rowsAffected int64
 	var err error
 	copyFn := func(conn *sql.Conn) (bool, error) {
-		rowsAffected, err = tdb.importBatch(conn, batch, args, exportDir)
+		rowsAffected, err = tdb.importBatch(conn, batch, args, exportDir, tableSchema)
 		return false, err
 	}
 	err = tdb.WithConn(copyFn)
@@ -546,7 +364,7 @@ func (tdb *TargetOracleDB) WithConn(fn func(*sql.Conn) (bool, error)) error {
 	return err
 }
 
-func (tdb *TargetOracleDB) importBatch(conn *sql.Conn, batch Batch, args *ImportBatchArgs, exportDir string) (rowsAffected int64, err error) {
+func (tdb *TargetOracleDB) importBatch(conn *sql.Conn, batch Batch, args *ImportBatchArgs, exportDir string, tableSchema map[string]map[string]string) (rowsAffected int64, err error) {
 	var file *os.File
 	file, err = batch.Open()
 	if err != nil {
@@ -591,14 +409,14 @@ func (tdb *TargetOracleDB) importBatch(conn *sql.Conn, batch Batch, args *Import
 
 	tableName := batch.GetTableName()
 	//TODO: figure out if directly the datatypes from schema files can be used instead of querying the db
-	if TableToColumnDataTypesCache[tableName] == nil {
-		TableToColumnDataTypesCache[tableName], err = tdb.getColumnDataTypes(tableName)
-		if err != nil {
-			return 0, fmt.Errorf("get column data types for table %s: %w", tableName, err)
-		}
-	}
-	columnTypes := TableToColumnDataTypesCache[tableName]
-	sqlldrConfig := args.GetSqlLdrControlFile(tdb.tconf.Schema, columnTypes)
+	// if TableToColumnDataTypesCache[tableName] == nil {
+	// 	TableToColumnDataTypesCache[tableName], err = tdb.getColumnDataTypes(tableName)
+	// 	if err != nil {
+	// 		return 0, fmt.Errorf("get column data types for table %s: %w", tableName, err)
+	// 	}
+	// }
+	// columnTypes := TableToColumnDataTypesCache[tableName]
+	sqlldrConfig := args.GetSqlLdrControlFile(tdb.tconf.Schema, tableSchema)
 	fileName := filepath.Base(batch.GetFilePath())
 
 	err = sqlldr.CreateSqlldrDir(exportDir)
@@ -771,7 +589,8 @@ func (tdb *TargetOracleDB) InitConnPool() error {
 	return nil
 }
 
-func (tdb *TargetOracleDB) GetDebeziumValueConverterSuite() map[string]ConverterFn {
+func (tdb *TargetOracleDB) GetDebeziumValueConverterSuite() map[string]tgtdbsuite.ConverterFn {
+	oraValueConverterSuite := tgtdbsuite.OraValueConverterSuite
 	for _, i := range []int{1, 2, 3, 4, 5, 6, 7, 8, 9} {
 		intervalType := fmt.Sprintf("INTERVAL YEAR(%d) TO MONTH", i) //for all interval year to month types with precision
 		oraValueConverterSuite[intervalType] = oraValueConverterSuite["INTERVAL YEAR TO MONTH"]
